+++ Runtime/Scripts/Interfaces/IConfirmView.cs
@@ -8,16 +8,21 @@ namespace ParkMinPackages.Workflow.Default.Interfaces
 	public interface IConfirmView
 	{
 		public UniTask ConfirmAsync(CancellationToken cancellationToken);
+	}
 
-		public static UniTask ConfirmAsyncByButton(
-			Button confirmButton,
+	public interface IButtonConfirmView : IConfirmView
+	{
+		public Button ConfirmButton { get; }
+
+		public static UniTask ConfirmAsync(
+			IButtonConfirmView view,
 			CancellationToken cancellationToken
 		) {
 			if (cancellationToken == CancellationToken.None) {
 				throw new ArgumentException("CancellationToken.None is not allowed", nameof(cancellationToken));
 			}
 
-			return confirmButton.OnClickAsync(cancellationToken);
+			return view.ConfirmButton.OnClickAsync(cancellationToken);
 		}
 	}
 }